backend/requirements.txt
# Core
fastapi==0.115.*
uvicorn[standard]==0.34.*
python-multipart==0.0.*
pydantic-settings==2.7.*

# LangChain + RAG
langchain==0.3.*
langchain-google-genai==2.*
langchain-chroma==0.2.*

# Vector store
chromadb==0.6.*

# Document parsing
pymupdf==1.25.*
python-docx==1.1.*
openpyxl==3.1.*

# Utilities
python-dotenv==1.1.*

# Testing
pytest==8.*
httpx==0.28.*
